Reorganise training code
