+++ grace/training/run.py
@@ -7,12 +7,12 @@ import torch
 from datetime import datetime
 from tqdm.auto import tqdm
 
-from grace.config import write_config_file, load_config_params
 from grace.io.image_dataset import ImageGraphDataset
 from grace.models.train import train_model
 from grace.models.datasets import dataset_from_graph
 from grace.models.classifier import GCN
 from grace.models.feature_extractor import FeatureExtractor
+from grace.training.config import write_config_file, load_config_params
 from grace.utils.transforms import get_transforms
